[JON-2] Get vehicles from json (#52)
* WIP, start adding boilerplate code

* delivery works

* changed root package name

* added domain tests

* added dto for repo and tests for repo

* added api tests

* tests\

* fixed annoying coveo format failing

* WIP, add observer transaction complete

* clean plus new tests

* Add observable tests

* Minimum for tests to pass and refactor

* cleanup

* reformat

* refactored comments

* pr refactor

* more refactoring

* refactor delivery test

* start of model refactor

* compiling + all tests pass

* finished refactor

* invalid model instead of invalid format postman query for vehicle

* merge fuckups by git

* removed .factory weird file

* really merge

* vehicle unit tests

* fix-request-order

* refactor and add test for addBattery

* InMemoryTransactionRepositoryTest

* Remove api exception in domain

* Vehicle tests

* use json property

* fix typo in models.json

* ModelDtoAssembler

* Delivery test.

* update pom

* model repo tests

* Battery assembler tests

* ModelDtoAssembler tests.

* add missing vehicle test

* remove useless getters

* Rename mappers

* fix typo

+++ src/main/java/ca/ulaval/glo4003/ws/api/transaction/dto/AddedBatteryResponse.java
@@ -1,13 +1,14 @@
 package ca.ulaval.glo4003.ws.api.transaction.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.math.BigDecimal;
 
 public class AddedBatteryResponse {
 
   @JsonProperty("estimated_range")
-  public Integer estimatedRange;
+  public BigDecimal estimatedRange;
 
-  public AddedBatteryResponse(Integer estimatedRange) {
+  public AddedBatteryResponse(BigDecimal estimatedRange) {
     this.estimatedRange = estimatedRange;
   }
 }
